requirements.txt
streamlit==1.28.1
PyMuPDF==1.23.8
pandas==2.1.3
openpyxl==3.11.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
PyInstaller==6.1.0
pydantic==2.5.0
